+++ src/qudi/gui/main_gui/main_gui.py
@@ -112,6 +112,9 @@ class QudiMainGui(GuiBase):
         self.mw.action_load_all_modules.triggered.connect(
             qudi_main.module_manager.activate_all_modules
         )
+        self.mw.action_deactivate_all_modules.triggered.connect(
+            qudi_main.module_manager.deactivate_all_modules
+        )
         self.mw.action_clear_all_appdata.triggered.connect(
             qudi_main.module_manager.clear_all_appdata
         )
@@ -140,6 +143,7 @@ class QudiMainGui(GuiBase):
         self.mw.action_reload_qudi.triggered.disconnect()
         self.mw.action_open_configuration_editor.triggered.disconnect()
         self.mw.action_load_all_modules.triggered.disconnect()
+        self.mw.action_deactivate_all_modules.triggered.disconnect()
         self.mw.action_clear_all_appdata.triggered.disconnect()
         self.mw.action_view_default.triggered.disconnect()
         # Disconnect signals from manager