+++ src/qudi/gui/main_gui/mainwindow.py
@@ -73,8 +73,15 @@ class QudiMainWindow(QtWidgets.QMainWindow):
         self.action_load_all_modules = QtWidgets.QAction()
         self.action_load_all_modules.setIcon(
             QtGui.QIcon(os.path.join(icon_path, 'dialog-warning')))
-        self.action_load_all_modules.setText('Load all modules')
-        self.action_load_all_modules.setToolTip('Load all available modules found in configuration')
+        self.action_load_all_modules.setText('Activate all modules')
+        self.action_load_all_modules.setToolTip(
+            'Activate all available modules found in configuration'
+        )
+        self.action_deactivate_all_modules = QtWidgets.QAction()
+        self.action_deactivate_all_modules.setIcon(
+            QtGui.QIcon(os.path.join(icon_path, 'edit-delete')))
+        self.action_deactivate_all_modules.setText('Deactivate all modules')
+        self.action_deactivate_all_modules.setToolTip('Deactivate all modules')
         self.action_clear_all_appdata = QtWidgets.QAction()
         self.action_clear_all_appdata.setIcon(QtGui.QIcon(os.path.join(icon_path, 'edit-clear')))
         self.action_clear_all_appdata.setText('Clear all AppData')
@@ -139,6 +146,7 @@ class QudiMainWindow(QtWidgets.QMainWindow):
         self.toolbar.addAction(self.action_reload_qudi)
         self.toolbar.addSeparator()
         self.toolbar.addAction(self.action_load_all_modules)
+        self.toolbar.addAction(self.action_deactivate_all_modules)
         self.toolbar.addAction(self.action_clear_all_appdata)
         self.addToolBar(self.toolbar)
 
@@ -151,6 +159,7 @@ class QudiMainWindow(QtWidgets.QMainWindow):
         menu.addAction(self.action_reload_qudi)
         menu.addSeparator()
         menu.addAction(self.action_load_all_modules)
+        menu.addAction(self.action_deactivate_all_modules)
         menu.addAction(self.action_clear_all_appdata)
         menu.addSeparator()
         menu.addAction(self.action_settings)